+++ examples/performance/src/Item.tsx
@@ -1,11 +1,11 @@
 import { useItem } from "dnd-timeline";
-import type { Relevance } from "dnd-timeline";
+import type { Span } from "dnd-timeline";
 import type { CSSProperties } from "react";
 import type React from "react";
 
 interface ItemProps {
 	id: string;
-	relevance: Relevance;
+	span: Span;
 	children: React.ReactNode;
 }
 
@@ -13,7 +13,7 @@ function Item(props: ItemProps) {
 	const { setNodeRef, attributes, listeners, itemStyle, itemContentStyle } =
 		useItem({
 			id: props.id,
-			relevance: props.relevance,
+			span: props.span,
 		});
 
 	const style: CSSProperties = {